requirements.txt
logzero
pandas
flask
smartapi-python
requests
flask-cors
supabase
pyotp
python-dotenv
gunicorn
websockets
websocket
yfinance